# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Order & Returns Management System API
  description: RESTful API for managing orders and returns for ArtiCurated marketplace
  version: 1.0.0
  contact:
    name: ArtiCurated API Support
    email: support@articurated.com

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Authentication
    description: User authentication and registration
  - name: Orders
    description: Order management operations
  - name: Returns
    description: Return request management operations
  - name: Payment Gateway
    description: Mock payment gateway endpoints

security:
  - bearerAuth: []

paths:
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new customer
      description: Creates a new customer account with CUSTOMER role
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request - validation error
        '409':
          description: Email already exists

  /api/v1/auth/register/admin:
    post:
      tags:
        - Authentication
      summary: Register an admin user
      description: Creates a new admin account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/v1/auth/register/manager:
    post:
      tags:
        - Authentication
      summary: Register a manager user
      description: Creates a new manager account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login and get JWT token
      description: Authenticates user and returns JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /api/v1/orders:
    post:
      tags:
        - Orders
      summary: Create a new order
      description: Creates a new order with PENDING_PAYMENT status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized

    get:
      tags:
        - Orders
      summary: Get orders by customer
      description: Retrieves all orders for a specific customer
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieves order details by order ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found

  /api/v1/orders/order-number/{orderNumber}:
    get:
      tags:
        - Orders
      summary: Get order by order number
      description: Retrieves order details by order number
      security:
        - bearerAuth: []
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

  /api/v1/orders/{id}/status:
    put:
      tags:
        - Orders
      summary: Update order status
      description: Updates order status (ADMIN only). Validates state transitions.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid state transition
        '403':
          description: Forbidden - Admin role required

  /api/v1/orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel an order
      description: Cancels an order (CUSTOMER or ADMIN). Only allowed from PENDING_PAYMENT or PAID status.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: reason
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Order cannot be cancelled in current status

  /api/v1/orders/{id}/history:
    get:
      tags:
        - Orders
      summary: Get order state history
      description: Retrieves complete audit trail of state changes for an order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: State history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateHistoryResponse'

  /api/v1/returns/order/{orderId}:
    post:
      tags:
        - Returns
      summary: Create a return request
      description: Creates a return request for a delivered order. Validates 7-day window and order status.
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReturnRequest'
      responses:
        '201':
          description: Return request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
        '400':
          description: Invalid return request (order not delivered, outside 7-day window, or active return exists)

  /api/v1/returns/{id}:
    get:
      tags:
        - Returns
      summary: Get return by ID
      description: Retrieves return request details by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Return details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'

  /api/v1/returns/{id}/status:
    put:
      tags:
        - Returns
      summary: Update return status
      description: Updates return status (ADMIN or MANAGER). Validates state transitions.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReturnStatusRequest'
      responses:
        '200':
          description: Return status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
        '400':
          description: Invalid state transition

  /api/v1/returns/{id}/history:
    get:
      tags:
        - Returns
      summary: Get return state history
      description: Retrieves complete audit trail of state changes for a return
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: State history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateHistoryResponse'

  /mock-payment-gateway/refund:
    post:
      tags:
        - Payment Gateway
      summary: Process refund
      description: Mock payment gateway endpoint for processing refunds
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          format: password
          example: password123

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzUxMiJ9...
        tokenType:
          type: string
          example: Bearer
        email:
          type: string
          example: customer@example.com
        role:
          type: string
          enum: [CUSTOMER, ADMIN, MANAGER]
          example: CUSTOMER

    OrderItemDto:
      type: object
      required:
        - productId
        - productName
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
          example: PROD-12345
        productName:
          type: string
          example: Handcrafted Ceramic Vase
        quantity:
          type: integer
          minimum: 1
          example: 2
        unitPrice:
          type: number
          format: decimal
          example: 99.99

    CreateOrderRequest:
      type: object
      required:
        - orderItems
        - shippingAddress
        - paymentMethod
        - paymentTransactionId
      properties:
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDto'
        shippingAddress:
          type: string
          example: 123 Main St, City, State 12345
        paymentMethod:
          type: string
          example: CREDIT_CARD
        paymentTransactionId:
          type: string
          example: TXN-123456789

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        orderNumber:
          type: string
          example: ORD-ABCD1234
        customerId:
          type: integer
          example: 1
        status:
          type: string
          enum: [PENDING_PAYMENT, PAID, PROCESSING_IN_WAREHOUSE, SHIPPED, DELIVERED, CANCELLED]
          example: PENDING_PAYMENT
        totalAmount:
          type: number
          format: decimal
          example: 199.98
        shippingAddress:
          type: string
        paymentMethod:
          type: string
        paymentTransactionId:
          type: string
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: decimal
        totalPrice:
          type: number
          format: decimal

    UpdateOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [PENDING_PAYMENT, PAID, PROCESSING_IN_WAREHOUSE, SHIPPED, DELIVERED, CANCELLED]
          example: PAID
        reason:
          type: string
          example: Payment confirmed

    CreateReturnRequest:
      type: object
      required:
        - returnReason
      properties:
        returnReason:
          type: string
          example: Product damaged during shipping

    ReturnResponse:
      type: object
      properties:
        id:
          type: integer
        returnNumber:
          type: string
          example: RET-ABCD1234
        orderId:
          type: integer
        status:
          type: string
          enum: [REQUESTED, APPROVED, REJECTED, IN_TRANSIT, RECEIVED, COMPLETED]
          example: REQUESTED
        returnReason:
          type: string
        refundAmount:
          type: number
          format: decimal
        managerNotes:
          type: string
        requestedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    UpdateReturnStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [REQUESTED, APPROVED, REJECTED, IN_TRANSIT, RECEIVED, COMPLETED]
          example: APPROVED
        managerNotes:
          type: string
          example: Return approved, customer can ship back

    StateHistoryResponse:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
          enum: [ORDER, RETURN]
        entityId:
          type: integer
        previousStatus:
          type: string
        newStatus:
          type: string
        changedBy:
          type: string
        changeReason:
          type: string
        createdAt:
          type: string
          format: date-time

    RefundRequest:
      type: object
      properties:
        orderId:
          type: integer
        returnId:
          type: integer
        amount:
          type: number
          format: decimal
        paymentTransactionId:
          type: string

    RefundResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        refundId:
          type: string
          example: REF-1234567890
        message:
          type: string
          example: Refund processed successfully
        amount:
          type: number
          format: decimal

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string